Swap contents of example.scala and PerfectNumbers.scala. Add Enumerations to config.json
